% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{refine_ordinate}
\alias{refine_ordinate}
\title{Knot Insertion, Removal, and Reinsertion}
\usage{
refine_ordinate(x, xi, theta, k = 4L)
}
\arguments{
\item{x}{the value of the knot to be inserted into the knot vector}

\item{xi}{the (whole) knot vector, including the repeated boundary knots}

\item{theta}{the ordinates of the control polygon vertices}

\item{order}{the order of the B-spline, defaults to 4 for cubic splines}
}
\value{
numeric vectors
}
\description{
Functions for the insertion, removal, and reinsertion of internal knots for
B-splines.
}

