% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bs.R
\name{bs}
\alias{bs}
\title{B-Splines}
\usage{
bs(x, ikntos, bknots, order = 4)
}
\arguments{
\item{x}{data}

\item{bknots}{boundary knot locations}

\item{order}{order of the piecewise polynomials}

\item{iknots}{internal knots}
}
\description{
An implimentation of Carl de Boor's recursive algorimthm for building
B-splines.
}
\details{
The difference between this function and \code{splines::bs} come in the
attributes associated with the output and default options.  The
\code{cpr::bs} call is intended to simplify the work needed with respect to
the control polygon reduction.
}

