% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsplines.R
\name{bsplineD1}
\alias{bsplineD1}
\alias{bsplineD1.cpr_bs}
\alias{bsplineD1.default}
\alias{bsplineD2}
\alias{bsplineD2.cpr_bs}
\alias{bsplineD2.default}
\title{B-spline Derivatives}
\usage{
bsplineD1(x, ...)

bsplineD2(x, ...)

\method{bsplineD1}{default}(x, iknots = numeric(0), bknots = range(x),
  order = 4L)

\method{bsplineD1}{cpr_bs}(x, ...)

\method{bsplineD2}{default}(x, iknots = numeric(0), bknots = range(x),
  order = 4L)

\method{bsplineD2}{cpr_bs}(x, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{iknots}{internal knots}

\item{bknots}{boundary knot locations, defaults to \code{range(x)}.}

\item{order}{order of the piecewise polynomials, defualts to 4L.}

\item{df}{degrees of freedom: sum of the order and internal knots.  Ignored
if \code{iknots} is specified.}

\item{x}{a \code{cpr_bs} object}

\item{x}{a \code{cpr_bs} object}
}
\description{
Generate the first and second derivatives of a B-spline Basis.
}
\examples{

}
\references{
C. de Boor, "A practical guide to splines. Revised Edition," Springer, 2001.
H. Prautzsch, W. Boehm, M. Paluszny, "Bezier and B-spline Techniques," Springer, 2002.
}

