% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tensor.R
\name{build_tensor}
\alias{build_tensor}
\title{Build Tensor}
\usage{
build_tensor(x, ...)
}
\arguments{
\item{x}{a matrix, or list of matrices, build the TP}

\item{...}{Matrices to build the TP}
}
\value{
A matrix
}
\description{
Tensor products of Matrices.
}
\examples{
A <- matrix(1:4, nrow = 10, ncol = 20)
B <- matrix(1:6, nrow = 10, ncol = 6)
build_tensor(A, B)
build_tensor(list(A, B))
build_tensor(A, B, B)

}
\author{
Peter DeWitt \email{dewittpe@gmail.com}
}
