% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn.R
\name{cn}
\alias{cn}
\alias{cn.cpr_bt}
\alias{cn.formula}
\alias{plot.cpr_cn}
\alias{print.cpr_cn}
\title{Control Nets}
\usage{
cn(x, ...)

\method{cn}{cpr_bt}(x, theta, ...)

\method{cn}{formula}(formula, data = parent.frame(), method = stats::lm,
  ...)

\method{print}{cpr_cn}(x, ...)

\method{plot}{cpr_cn}(x, ..., show_spline = FALSE, color = FALSE, n = 100)
}
\arguments{
\item{x}{a \code{cpr_bs} object}

\item{...}{arguments passed to the regression method}

\item{theta}{a vector of (regression) coefficients, the ordinates of the
control net.}

\item{formula}{a formula that is appropriate for regression method being
used.}

\item{data}{see documentation in \code{\link[stats]{lm}}}

\item{method}{the regression method such as \code{\link[stats]{lm}},
\code{\link[stats]{glm}}, \code{\link[lme4]{lmer}}, \code{\link[geepack]{geeglm}}, ...}

\item{show_spline}{boolean (default FALSE) to plot the spline function within
its control net}

\item{color}{boolean (default FALSE) if more than one \code{cpr_cn} object is
to be plotted, set this value to TRUE to have the graphic in color (linetypes
will be used regardless of the color settting).}

\item{n}{the number of data points to use for plotting the spline}
}
\description{
Generate the control net for a uni-variable B-spline
}
\details{
\code{cn} generates the control net for the given B-spline function.

\code{cpr} runs the control net reduction algorithm
}

