% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr_diagnostics.R
\name{plot.cpr_cnr}
\alias{plot.cpr_cnr}
\title{Control Net Reduction Plots}
\usage{
\method{plot}{cpr_cnr}(x, type = "rmse", from = 1, to, ...)
}
\arguments{
\item{x}{a \code{cpr_cnr} object}

\item{type}{type of diagnostic plot.
\code{"loglik"} for the loglikihood by degrees of freedom,
\code{"rmse"} for root mean squared residuals by model index}

\item{from}{the first index of \code{x} to plot}

\item{to}{the last index of \code{x} to plot}

\item{...}{ignored}
}
\description{
A collection of function for the inspection and evaluation of the control
polygon reduction.
}
\author{
Peter DeWitt \email{dewittpe@gmail.com}
}
