% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_diagnostics.R
\name{cp_value}
\alias{cp_diff}
\alias{cp_value}
\title{Control Polygon Diagnostics}
\usage{
cp_value(obj, x)

cp_diff(cp1, cp2, denom = diff(range(cp2$theta)))
}
\arguments{
\item{obj}{a cpr_cp object or \code{data.frame} where the first column is the
abscissa and the second column is the ordinate for the control polygon vertices.}

\item{x}{absicissa at which to determine the ordinate on control polygon cp}

\item{denom}{the denominator used for relative distance between the control
vertices of cp1 to control polygon cp2.  setting \code{denom = 1} will return
the absolute distance.  The default \code{diff(range(cp2$theta))} scales the}
}
\value{
\code{cp_value} returns the ordinate on the control polygon line segment for
the absicissa \code{x} given.  \code{x} could be a control vertex or on a
line segement defined by two control vertices of the control polygon
provided.

\code{cp_diff} returns the (relative) distance between the control vertices
of cp1 to the control polygons cp2.
}
\description{
Collection of functions for insepction and analysis of the control polygons
}

