% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{cpr-defunct}
\alias{cpr-defunct}
\alias{refine_ordinate}
\alias{coarsen_ordinate}
\alias{hat_ordinate}
\alias{insertion_matrix}
\alias{wiegh_iknots}
\alias{influence_of}
\alias{influence_weights}
\title{Defunct Functions}
\usage{
refine_ordinate(...)

coarsen_ordinate(...)

hat_ordinate(...)

insertion_matrix(...)

wiegh_iknots(...)

influence_of(...)

influence_weights(...)
}
\arguments{
\item{...}{pass through}
}
\description{
A major refactor of the package between v0.3.0 and v.0.4.0 took place and
many functions were made defunct.  The refactor was so extensive that moving
the functions to deprecated was not a viable option.
}
