% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr-package.R
\docType{package}
\name{cpr-package}
\alias{cpr-package}
\title{cpr: Control Polygon Reduction}
\description{
The cpr package implements the control polygon reduction method presented in
MY TO BE PUBLISHED paper(s).

There are several options that the user can set while working with this
package that will help with the reporting and plotting of results.

\itemize{
   \item Options...  Define them
}
}

