% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_polygon_reduction.R
\name{cpr}
\alias{cpr}
\title{Control Polygon Reduction}
\usage{
cpr(formula, data = parent.env(), method = lm, p = 2L, ...)
}
\arguments{
\item{formula}{a formula that is appropriate for regression method being
used.}

\item{data}{see documentation in \code{\link[stats]{lm}}}

\item{method}{the regression method such as \code{\link[stats]{lm}},
\code{\link[stats]{glm}}, \code{\link[lme4]{lmer}}, \code{\link[geepack]{geeglm}}, ...}

\item{p}{defaults to 2L, the L^p norm used in determining the 'weight of
importance' of each internal knot.}

\item{...}{arguments passed to the regression method}
}
\description{
Run the Control Polygon Reduction Algorithm.
}
\details{
\code{cpr} runs the control polygon reduction algorithm
}

