% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr_functions.R
\name{cpr2BD}
\alias{cpr2BD}
\title{This function returns a birthday based on a real cpr-number}
\usage{
cpr2BD(cpr)
}
\arguments{
\item{cpr}{a character vector representating the cpr-numbers. must be 10 digits long and containg no '-'.}
}
\value{
a date vector of birthdates of the individuals.
}
\description{
This function returns a birthday based on a real cpr-number
}
\details{
Based on the digits 1-2 day of birth is calculated, the month is calculated based on digits 3-4 and year is
calculated based on digits 5-7.
}
\seealso{
\code{cpr_correct} \code{cpr2Sex}
}
\author{
Magnus Ahlström
}
