% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr_functions.R
\name{cpr2Sex}
\alias{cpr2Sex}
\title{This function returns sex based on a real cpr-number}
\usage{
cpr2Sex(cpr, output = c("as.numeric", "factor"))
}
\arguments{
\item{cpr}{a character vector representating the cpr-numbers. must be 10 digits long and containg no '-'.}

\item{output}{class of the output either 'numeric' 1 is male and 2 is female. Or factor with labels
'M', 'F' or 'Invalid'}
}
\value{
a date vector of birthdates of the individuals.
}
\description{
This function returns sex based on a real cpr-number
}
\details{
Based on digit 10 sex is calculated, if the number is dividable by 2 output is female if not then
output is male.
}
\seealso{
\code{cpr_correct} \code{cpr2BD}
}
\author{
Magnus Ahlström
}
