% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr_functions.R
\name{cpr_correct}
\alias{cpr_correct}
\title{This function checks if the cpr-number is a real cpr-number.}
\usage{
cpr_correct(cpr)
}
\arguments{
\item{cpr}{a character vector representating the cpr-numbers. must be 10 digits long and containg no '-'.}
}
\value{
logical indicating whether the cpr-number is real or not
}
\description{
This function checks if the cpr-number is a real cpr-number.
}
\details{
A real cpr number contains numbers from 0-9 is 10 digits long and when the digits are multiplied by
a specific vector, the rowsum should be dividable by 11.
}
\seealso{
\code{cpr2BD} \code{cpr2Sex}
}
\author{
Magnus Ahlström
}
