% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cpr_bsplines.R
\name{extract_cpr_bsplines}
\alias{extract_cpr_bsplines}
\title{Extract the bspline or btensor call from a formula}
\usage{
extract_cpr_bsplines(form)
}
\arguments{
\item{form}{a formula}
}
\description{
Non-exported function.  Might be depericated.
}
\author{
Peter DeWitt \email{dewittpe@gmail.com}
}
