% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spline.R
\name{get_spline}
\alias{get_spline}
\title{Get Spline Function}
\usage{
get_spline(x, margin, at, n = 100)
}
\arguments{
\item{x}{a \code{cpr_cp} or \code{cpr_cn} object.}

\item{margin}{an integer vector or length 1 or 2.  Only used when working
x is a \code{cpr_cn} object.  #'}

\item{at}{point value for marginals not defined in the \code{margin}.  Only
used when \code{x} is a \code{cpr_cn} object.}

\item{n}{the length of sequence to use for interpolating the spline function.}
}
\description{
Generate a \code{data.frame} for interpolating and plotting a spline
function, given a \code{cpr_cp} or \code{cpr_cn} object.
}

