% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.cpr_bs}
\alias{is.cpr_bs}
\alias{is.cpr_bt}
\alias{is.cpr_cn}
\alias{is.cpr_cnr}
\alias{is.cpr_cp}
\alias{is.cpr_cpr}
\title{Tests for Objects}
\usage{
is.cpr_bs(x)

is.cpr_bt(x)

is.cpr_cp(x)

is.cpr_cpr(x)

is.cpr_cn(x)

is.cpr_cnr(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Test if an R object is of a class specific to the cpr package
}
\author{
Peter DeWitt \email{dewittpe@gmail.com}
}

