% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knot_expr.R
\name{knot_expr}
\alias{knot_expr}
\title{Knot Expressions}
\usage{
knot_expr(x, digits)
}
\arguments{
\item{x}{a \code{cpr_cp} or \code{cpr_bs} object}

\item{digits}{digits to the right of the decimal point to report}
}
\description{
Non-exported function used to build expressions for the knot sequences to be
labeled well on a plot.
}
