% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newknots.R
\name{newknots}
\alias{newknots}
\title{New Knots for CPs and CNs in CPR and CNR}
\usage{
newknots(form, nk)
}
\arguments{
\item{nk}{}
}
\description{
Non-exported function, \code{newknots} are used in the cpr and cnr calls.
}
