% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn.R
\name{plot.cpr_cn}
\alias{plot.cpr_cn}
\title{Plotting Control Nets}
\usage{
\method{plot}{cpr_cn}(x, margin = 1:2, at, ...)
}
\arguments{
\item{x}{a \code{cpr_cn} object}

\item{margin}{an integer vector or length 1 or 2.  For length 1, the marginal
control polygon is plotted via a call to \code{\link{plot.cpr_cp}}.  For
length 2, the (marginal) tensor surface is plotted.  See details.}

\item{at}{point value for marginals not defined in the \code{margin}.  See
details.}

\item{\ldots}{arguments passed to \code{\link{plot.cpr_cp}}}
}
\description{
One- and Two-dimensional plots of control nets.
}

