% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cpr_cpr.R
\name{plot.cpr_cpr}
\alias{plot.cpr_cpr}
\title{Control Polygon Reduction Plots}
\usage{
\method{plot}{cpr_cpr}(x, type = c("cps", "loglik", "rss", "rse"), from = 1, to, ...)
}
\arguments{
\item{x}{a \code{cpr_cpr} object}

\item{type}{type of diagnostic plot.  \code{"cps"} for control polygons,
\code{"loglik"} for the log likelihood by degrees of freedom,
\code{"rse"} for residual standard error by index.}

\item{from}{the first index of \code{x} to plot}

\item{to}{the last index of \code{x} to plot}

\item{...}{arguments passed to \code{plot.cpr_cp}}
}
\description{
A wrapper around several ggplot2 calls to help evaluate results of a CPR run.
}
\seealso{
\code{\link{plot.cpr_cp}}, \code{\link{cpr}}, \code{\link{cp}}
}
