% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_of.R
\name{plot.cpr_influence_of}
\alias{plot.cpr_influence_of}
\title{Plotting for cpr_influence_of objects}
\usage{
\method{plot}{cpr_influence_of}(x, ...)
}
\arguments{
\item{x}{a \code{cpr_influence_of} object}

\item{...}{Arguments passed to \code{\link{plot.cpr_cp}}}
}
\value{
a \code{ggplot2} graphic
}
\description{
Plot method for cpr_influence_of objects
}
\seealso{
\code{\link{influence_of}}
}
