% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsplines.R
\name{print.cpr_bs}
\alias{print.cpr_bs}
\title{Print bsplines}
\usage{
\method{print}{cpr_bs}(x, n = 6L, ...)
}
\arguments{
\item{x}{a \code{cpr_bs} object.}

\item{n, }{number of rows of the B-spline basis matrix to display, defaults to
6L.}

\item{\ldots}{not currently used.}
}
\description{
Print bsplines
}
