% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cpr_cn.R
\name{summary.cpr_cn}
\alias{summary.cpr_cn}
\title{Summary of Control Net}
\usage{
\method{summary}{cpr_cn}(object, ...)
}
\arguments{
\item{object}{a \code{cpr_cn} object}

\item{...}{pass through}
}
\value{
a \code{data.frame}
}
\description{
Generate a summary of control net object
}
\examples{

acn <- cn(log10(pdg) ~ btensor(list(day, age)
                               , df = list(10, 8)
                               , bknots = list(c(-1, 1), c(44, 53)))
         , data = spdg)

summary(acn)

}
