% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cpr_cnr.R
\name{summary.cpr_cnr}
\alias{summary.cpr_cnr}
\title{Summarize Control Net Reduction Objects}
\usage{
\method{summary}{cpr_cnr}(object, ...)
}
\arguments{
\item{object}{a \code{cpr_cnr} object}

\item{...}{pass through}
}
\value{
a \code{cpr_summary_cpr_cnr} object, that is just a \code{data.frame}
}
\description{
Summarize Control Net Reduction Objects
}
\examples{

acn <- cn(log10(pdg) ~ btensor(list(day, age)
                               , df = list(10, 8)
                               , bknots = list(c(-1, 1), c(44, 53)))
         , data = spdg)
cnr0 <- cnr(acn)
cnr0
summary(cnr0)

}
