% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cpr_cpr.R
\name{summary.cpr_cpr}
\alias{summary.cpr_cpr}
\title{Summarize a Control Polygon Reduction Object}
\usage{
\method{summary}{cpr_cpr}(object, ...)
}
\arguments{
\item{object}{a \code{cpr_cpr} object}

\item{....}{pass through}
}
\description{
Summarize a Control Polygon Reduction Object
}
