% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bsplines.R
\name{update_bsplines}
\alias{update_bsplines}
\alias{update_btensor}
\title{Update bspline or btensor calls}
\usage{
update_bsplines(object, ..., evaluate = TRUE)

update_btensor(object, ..., evaluate = TRUE)
}
\arguments{
\item{object}{an object to update.}

\item{...}{things to update, expected to be \code{iknots}, \code{df},
\code{bknots}, or \code{order}.}

\item{evaluate}{If true evaluate the new call else return the call.}
}
\description{
Update \code{cpr_bs} and \code{cpr_bt} objects alone or within \code{cpr_cp}
and \code{cpr_cn} objects.
}
\seealso{
\code{\link[stats]{update}}, \code{\link{bsplines}},
\code{\link{btensor}}
}

