% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{us_covid_cases}
\alias{us_covid_cases}
\title{United States Laboratoy Confirmed COVID-19 Cases}
\format{
a \code{data.frame} with two columns
\describe{
  \item{date}{year, month, day}
  \item{cases}{number of reported laboratory-confirmeded COVID-19 cases}
}
}
\source{
Download original data from
<https://data.cdc.gov/Case-Surveillance/COVID-19-Case-Surveillance-Public-Use-Data/vbim-akqf>
on December 5, 2023.  The reported data set was last updated on November 3,
2023.
}
\usage{
us_covid_cases
}
\description{
Number of laboratory-confirmed COVID-19 cases in the United States, as
reported by the Centers for Disease Control, between January 1 2020 and May
11, 2023, the end of the public health emergency declaration.
}
\keyword{datasets}
