% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/bsplines.R
\name{greville_sites}
\alias{bsplines}
\alias{greville_sites}
\alias{is.cpr_bs}
\title{B-Splines}
\usage{
greville_sites(xi, order)

bsplines(x, iknots = NULL, df = NULL, bknots = range(x), order = 4L)

is.cpr_bs(x)
}
\arguments{
\item{xi}{full knot vector}

\item{order}{order of the piecewise polynomials, defualts to 4L.}

\item{x}{a numeric vector}

\item{iknots}{internal knots}

\item{bknots}{boundary knot locations, defaults to \code{range(x)}.}
}
\description{
An implimentation of Carl de Boor's recursive algorimthm for building
B-splines.
}
\details{
The difference between this function and \code{splines::bs} come in the
attributes associated with the output and default options.  The
\code{cpr::bs} call is intended to simplify the work needed with respect to
the control polygon reduction.

\code{greville_sites} returns the knot averages.
}
\examples{
# build a vector of values to transform
xvec <- seq(-3, 5, length = 100)

# cubic b-spline
bmat <- bsplines(xvec, iknots = c(-2, 0, 0.2))
bmat

# view the splines
plot(bmat)

# If you want a second x-axis to show the x-values try the following:
second_x_axis <- round(stats::quantile(xvec, probs = seq(0, 1, by = .2)), 2)

plot(bmat) + 
ggplot2::annotate(geom = "text", x = second_x_axis, y = -0.02, label = second_x_axis) + 
ggplot2::annotate(geom = "linerange", x = second_x_axis, ymin = -0.05, ymax = -0.04) + 
ggplot2::coord_cartesian(ylim = c(0, 1))

# quadratic splines
bmat <- bsplines(xvec, iknots = c(-2, 0, 0.2), order = 3L)
bmat
plot(bmat) + ggplot2::ggtitle("Quadratic B-splines")

}

