% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/bsplines.R
\name{greville_sites}
\alias{bsplines}
\alias{greville_sites}
\alias{is.cpr_bs}
\title{B-Splines}
\usage{
greville_sites(xi, order)

bsplines(x, iknots = numeric(0), bknots = range(x), order = 4L)

is.cpr_bs(x)
}
\arguments{
\item{xi}{full knot vector}

\item{order}{order of the piecewise polynomials, defualts to 4L.}

\item{x}{a numeric vector}

\item{iknots}{internal knots}

\item{bknots}{boundary knot locations, defaults to \code{range(x)}.}
}
\description{
An implimentation of Carl de Boor's recursive algorimthm for building
B-splines.
}
\details{
The difference between this function and \code{splines::bs} come in the
attributes associated with the output and default options.  The
\code{cpr::bs} call is intended to simplify the work needed with respect to
the control polygon reduction.

\code{greville_sites} returns the knot averages.
}

