% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsplines.R
\name{bsplines}
\alias{bsplines}
\alias{greville_sites}
\alias{plot.cpr_bs}
\alias{print.cpr_bs}
\title{B-Splines}
\usage{
bsplines(x, iknots = numeric(0), bknots = range(x), order = 4L)

\method{print}{cpr_bs}(x, n = 6L, ...)

\method{plot}{cpr_bs}(x, ggplot2 = getOption("cpr_ggplot2", FALSE), ...)

greville_sites(xi, order = 4L)
}
\arguments{
\item{x}{data}

\item{iknots}{internal knots}

\item{bknots}{boundary knot locations, defaults to \code{range(x)}.}

\item{order}{order of the piecewise polynomials}

\item{n, }{number of rows of the B-spline basis matrix to display, defaults to
6L.}

\item{...}{passed to \code{graphics::matplot}}

\item{ggplot2}{boolean, if TRUE, return a list of layers to plot with
ggplot(), if false, pass x and ... to matplot}

\item{xi}{full knot vector,}
}
\description{
An implimentation of Carl de Boor's recursive algorimthm for building
B-splines.
}
\details{
The difference between this function and \code{splines::bs} come in the
attributes associated with the output and default options.  The
\code{cpr::bs} call is intended to simplify the work needed with respect to
the control polygon reduction.

\code{greville_sites} returns the knot averages.
}

