% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cp.R
\name{cp}
\alias{cp}
\alias{cpr}
\title{Control Polygons}
\usage{
cp(formula, data = parent.env(), method = lm, ...)

cpr(formula, data = parent.env(), method = lm, psi_f = 170, psi_n = 165,
  K = 0, ...)
}
\arguments{
\item{formula}{a formula that is appropriate for regression method being
used.}

\item{data}{see documentation in \code{\link{lm}}}

\item{method}{the regression method such as \code{\link{lm}},
\code{\link{glm}}, \code{\link{lmer}}, \code{\link{geeglm}}, ...}

\item{...}{arguments passed to the regression method}

\item{psi_f}{the limit for effective colinear fair points}

\item{psi_n}{the limit for effective colinear non-fair points}

\item{K}{reduce the control polygon, regardels of psi_f and psi_n untill K or
fewer internal knots are left.}
}
\description{
Generate the control polygon for a univariable B-spline
}
\details{
\code{cp} generates the control polygon for the given B-spline function.

\code{cpr} runs the control polygon reduction algorithm
}

