% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp.R
\name{cp}
\alias{cp}
\alias{cpr}
\title{Control Polygons}
\usage{
cp(formula, data = parent.env(), method = lm, ...)

cpr(formula, data = parent.env(), method = lm, p = 2L, ...)
}
\arguments{
\item{formula}{a formula that is appropriate for regression method being
used.}

\item{data}{see documentation in \code{\link[stats]{lm}}}

\item{method}{the regression method such as \code{\link[stats]{lm}},
\code{\link[stats]{glm}}, \code{\link[lme4]{lmer}}, \code{\link[geepack]{geeglm}}, ...}

\item{...}{arguments passed to the regression method}
}
\description{
Generate the control polygon for a univariable B-spline
}
\details{
\code{cp} generates the control polygon for the given B-spline function.

\code{cpr} runs the control polygon reduction algorithm
}

