% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr-package.R
\docType{package}
\name{cpr-package}
\alias{cpr-package}
\alias{_PACKAGE}
\title{cpr: Control Polygon Reduction}
\description{
The cpr package implements the control polygon reduction and control net
reduction methods for finding parsimonious B-spline regression models as
described in DeWitt (2017).
}
\references{
DeWitt, Peter Edward. \dQuote{Parsimonious B-spline regression models via
control polygon and control net reduction for identifying factors explaining
variation in daily hormone profiles during the menopausal transition.}
(2017).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dewittpe/cpr/}
  \item \url{http://www.peteredewitt.com/cpr/}
  \item Report bugs at \url{https://github.com/dewittpe/cpr/issues}
}

}
\author{
\strong{Maintainer}: Peter DeWitt \email{dewittpe@gmail.com} (\href{https://orcid.org/0000-0002-6391-0795}{ORCID})

Other contributors:
\itemize{
  \item Samantha MaWhinney \email{sam.mawhinney@ucdenver.edu} [thesis advisor]
  \item Nichole Carlson \email{nichole.carlson@ucdenver.edu} [thesis advisor]
}

}
\keyword{internal}
