% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr-package.R
\docType{package}
\name{cpr-package}
\alias{cpr-package}
\title{cpr: Control Polygon Reduction}
\description{
The cpr package implements the control polygon reduction and control net
reduction method presented in
MY TO BE PUBLISHED paper(s).

This package was part of Peter DeWitt's Ph.D. dissertation which was overseen
by his advisors Samantha MaWhinney and Nichole Carlson.
}

