% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpr.R
\docType{package}
\name{cpr-package}
\alias{cpr-package}
\title{cpr: Control Polygon Reduction}
\description{
The cpr package implements the control polygon reduction method presented in
MY TO BE PUBLISHED paper(s).
}

