% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_polygon_reduction.R
\name{cpr}
\alias{cpr}
\title{Control Polygon Reduction}
\usage{
cpr(formula, data = parent.env(), method = lm, p = 2L, keep = -1L, ...)
}
\arguments{
\item{formula}{a formula that is appropriate for regression method being
used.}

\item{data}{see documentation in \code{\link[stats]{lm}}}

\item{method}{the regression method such as \code{\link[stats]{lm}},
\code{\link[stats]{glm}}, \code{\link[lme4]{lmer}}, \code{\link[geepack]{geeglm}}, ...}

\item{p}{defaults to 2L, the L^p norm used in determining the influence
weight of each internal knot.}

\item{keep}{an integer value (defaults to -1L), the number of regression fits
to keep.  See Details.}

\item{...}{arguments passed to the regression method}
}
\description{
Run the Control Polygon Reduction Algorithm.
}
\details{
\code{cpr} runs the control polygon reduction algorithm.

\code{keep} will keep the regression fit as part of the \code{cpr\_cp} object
for models with upto and including keep fits.  For example, if \code{keep =
10} then the resulting \code{cpr\_cpr} object will have the regression fit
stored in the first \code{keep + 1} (zero internal knots, one internal knot,
\ldots, \code{keep} internal knots) \code{cpr\_cp} objects in the list.  The
limit on the number of stored regression fits is to keep memory usage down.
}

