% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_polygon_reduction.R
\name{cpr}
\alias{cpr}
\title{Control Polygon Reduction}
\usage{
cpr(x, p = 2, ...)
}
\arguments{
\item{x}{a \code{cpr_cp} or \code{cpr_tensor} object}

\item{p}{defaults to 2L, the L^p norm used in determining the influence
weight of each internal knot.}

\item{...}{not currently used}
}
\description{
Run the Control Polygon Reduction Algorithm.
}
\details{
\code{cpr} runs the control polygon reduction algorithm.

\code{keep} will keep the regression fit as part of the \code{cpr\_cp} object
for models with upto and including keep fits.  For example, if \code{keep =
10} then the resulting \code{cpr\_cpr} object will have the regression fit
stored in the first \code{keep + 1} (zero internal knots, one internal knot,
\ldots, \code{keep} internal knots) \code{cpr\_cp} objects in the list.  The
limit on the number of stored regression fits is to keep memory usage down.
}

