% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr_diagnostics.R
\name{cpr_select}
\alias{cpr_layers}
\alias{cpr_select}
\title{Control Polygon Reduction Diagnoses}
\usage{
cpr_select(obj, tol = 0.01)

cpr_layers(obj, start = 1L, end = 6L)
}
\arguments{
\item{obj}{a cpr_cpr object}

\item{tol}{max difference/error between the vertices of cp1 to cp2}

\item{start}{index of the first control polygon to plot}

\item{end}{index of the last control polygon to plot}
}
\description{
A collection of function for the inspection and evaluation of the control
polygon reduction.
}

