% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr_diagnostics.R
\name{plot.cpr_cpr}
\alias{plot.cpr_cpr}
\title{Control Polygon Reduction Plots}
\usage{
\method{plot}{cpr_cpr}(x, type = "cps", from = 1, to, ...)
}
\arguments{
\item{x}{a \code{cpr_cpr} object}

\item{type}{type of diagnostic plot.  \code{"cps"} for control polygons,
\code{"loglik"} for the loglikihood by degrees of freedom,
\code{"rmse"} for root mean squared residuals by degrees of freedom}

\item{from}{the first index of \code{x} to plot}

\item{to}{the last index of \code{x} to plot}

\item{...}{args passed to \code{cpr::plot.cpr_cp}}
}
\description{
A collection of function for the inspection and evaluation of the control
polygon reduction.
}
\seealso{
\code{\link{plot.cpr_cp}}
}
