% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iknots_or_df.R
\name{iknots_or_df}
\alias{iknots_or_df}
\title{Internal Knots or Degrees of Freedom}
\usage{
iknots_or_df(x, iknots, df, order)
}
\arguments{
\item{order}{}
}
\value{
a numeric vector to use as the internal knots defining a B-spline
}
\description{
Check order, degrees of freedom (df) and iknots
}
\details{
This is an internal function, not to be exported, and used in the calls for
\code{\link{bsplines}} and \code{\link{bsplineD}}.
}
\seealso{
\code{\link{bsplines}}, \code{\link{bsplineD}}
}
