% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iknots_or_df.R
\name{iknots_or_df}
\alias{iknots_or_df}
\title{Internal Knots or Degrees of Freedom}
\usage{
iknots_or_df(x, iknots, df, order)
}
\arguments{
\item{order}{}
}
\value{
a numeric vector to use as the internal knots defining a B-spline
}
\description{
Used in the \link{bsplines} call to check order, degrees of freedom (df) and
iknots
}
\seealso{
\link{bsplines}
}
