% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_weights.R
\name{influence_weights}
\alias{influence_weights}
\title{Influence Weights}
\usage{
influence_weights(x, p = 2, n_polycoef = 50L)
}
\arguments{
\item{x}{a \code{cpr_cp} or \code{cpr_tensor} object}

\item{p}{the order of the norm, default \code{p = 2}.}

\item{n_polycoef}{for influence weights in tensor products, this parameter
set the number of polynomial coefficients to use in each of the marginal
calculations.  Ignored for \code{cpr_cp} objects.}
}
\value{
A data_frame with two elements, the internal knots (iknots) and the weights.
}
\description{
Determine the influence weight of each iternal knot on each marginal of a
tensor product.
}
\details{
\code{parallel::mclapply} is used when generating the influence weights for a
\code{cpr_cn} object.  By default, \code{parallel::mclapply} uses two cores.
To use more than two cores, e.g., to use three cores, set
\code{options(mc.cores = 3L)}.  NOTE:  There is no benefit to using more
cores than there are margins of the tensor product.
}

