% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_weights.R
\name{influence_weights}
\alias{influence_weights}
\title{Influence Weights}
\usage{
influence_weights(tp, theta)
}
\arguments{
\item{tp}{a cpr_tensor object}

\item{theta}{corresponding ordinates for \code{x}}
}
\value{
A list with each element a numeric vector of influence weights for the
internal knots on a marginal
}
\description{
Determine the influence weight of each iternal knot on each marginal of a
tensor product.
}
\examples{
## NOT RUN

data("diamonds", package = "ggplot2")
sub_diamonds <- diamonds[1:11000, ]

tp_diamonds <- with(sub_diamonds, cpr::tensor(list(carat, depth, table), iknots = list(c(0.5, 1.5), numeric(0), numeric(0))))

fit <- lm(price ~ tp_diamonds + 0, data = sub_diamonds)
theta_diamonds <- unname(coef(fit))

influence_weights(tp_diamonds, theta_diamonds)


}

