% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsplines.R
\name{plot.cpr_bs}
\alias{plot.cpr_bs}
\title{Plot B-splines}
\usage{
\method{plot}{cpr_bs}(x, ...)
}
\arguments{
\item{x}{a \code{cpr_bs} object}

\item{\ldots}{not currently used}
}
\description{
Plot B-splines
}

