% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.cpr_cp}
\alias{predict.cpr_cp}
\title{Model Prediction}
\usage{
\method{predict}{cpr_cp}(object, newdata, ...)
}
\arguments{
\item{object}{a \code{cpr_cp} or \code{cpr_cn} object}

\item{newdata}{a \code{data.frame}}

\item{...}{passed to \code{\link[stats]{predict}} from the \code{stats}
package.}
}
\description{
Model prediction for \code{cpr_cp} and \code{cpr_cn} objects.
}

