% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle.R
\name{sign_changes}
\alias{sign_changes}
\title{Sign Changes}
\usage{
sign_changes(
  object,
  lower = min(object$bknots),
  upper = max(object$bknots),
  n = 1000,
  derivative = c(1L, 2L),
  ...
)
}
\arguments{
\item{object}{a \code{cpr_cp} object}

\item{lower}{the lower limit of the integral}

\item{upper}{the upper limit of the integral}

\item{n}{number of values to assess the derivative between \code{lower} and
\code{upper}.}

\item{derivative}{integer value denoted first or second derivative}

\item{...}{pass through}
}
\value{
the number of times the sign of the first or second derivative
changes within the specified interval.
}
\description{
Count the number of times the first, or second, derivative of a spline
changes sign.
}
\seealso{
\code{\link{wiggle}}
}
