% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cpr_cnr.R
\name{summary.cpr_cnr}
\alias{summary.cpr_cnr}
\title{Summarize Control Net Reduction Objects}
\usage{
\method{summary}{cpr_cnr}(object, ...)
}
\arguments{
\item{object}{a \code{cpr_cnr} object}

\item{...}{pass through}
}
\description{
Summarize Control Net Reduction Objects
}
