% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cpr_cp.R
\name{summary.cpr_cp}
\alias{summary.cpr_cp}
\title{Summarize a Control Polygon Object}
\usage{
\method{summary}{cpr_cp}(object, wiggle = TRUE, integrate.args = list(), ...)
}
\arguments{
\item{object}{a \code{cpr_cp} object}

\item{wiggle}{logical, if \code{TRUE} then the integral of the squared second
derivative of the spline function will be calculated via
\code{\link[stats]{integrate}}.}

\item{integrate.args}{a list of arguments passed to \code{\link{wiggle}} and
ultimately \code{\link[stats]{integrate}}.}
}
\description{
Summarize a Control Polygon Object
}
