% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle.R
\name{wiggle}
\alias{wiggle}
\title{Wiggliness of a Spline function}
\usage{
wiggle(object, lower, upper, ...)
}
\arguments{
\item{object}{a \code{cpr_cp} object}

\item{lower}{the lower limit of the intergral}

\item{upper}{the upper limit of the intergral}

\item{...}{arguments passed to \code{stats::integrate}}
}
\description{
Calculate the integral of the squared second derivate of the spline function.
}
\details{
\deqn{\int \left( \frac{\mathrm{d}^2}{\mathrm{d}x^2} f \left(x \right)
\right)^2 \mathrm{d}x.}{ \int (d^2 / dx^2 f(x))^2 dx}
}
\seealso{
\code{\link{cp}} \code{\link[stats]{integrate}}
}

